Made the day labels in agenda/week view use the same width. Following the instructions in the Qt docs for having many widgets in a QHBoxLayout with exactly the same with, I set the minimum size to 1 (so the window should no longer resized to a width larger than the screen, but instead always stay the same size) and set the stretch to 1, too. As a consequence, text that doesn't fit into the available space will be cut.
If this happens with the "Mon 11" labels, the date will not be visible, which I solved the following general way:

I also added a new class KOAlternativeLabel, which takes one short and one long string. If the long string is too long to be display in the label, the short string is used. I make use of this new function in the lay labels, where the long string is "Mon 11", and the short just "11". So if there is not enough space available in the header, I don't print the weekday any more.

CCMAIL: 54483-done@bugs.kde.org, 59580-done@bugs.kde.org

svn path=/trunk/kdepim/; revision=241579