Finally, korganizer supports overnight working hours, and allows the user to select which days are working days (using seven check boxes, one for each day of the week).
A few words concerning night shift:
For overnight working hours, I use the day when the shift starts to determine if the shift is on a holiday, and thus shouldn't be drawn as work shift. As an example: If you have working hours Mon-Fri from 20:00 - 05:00, then you'll see the following shifts:
Mon 20:00 - Tue 05:00
Tue 20:00 - Wed 05:00
...
Fri 20:00 - Sat 05:00

But NOT Sat 20:00-Sun05:00 and also not Sun 20:00 - Mon 05:00 (because in these cases the shift starts on a day that is not a working day). The same applies to holidays: If the tuesday is a holiday, the shift from tuesday evening to wednesday morning will not be shown. I hope this is the expected behavior...

CCMAIL: 70053-done@bugs.kde.org, 56611-done@bugs.kde.org

svn path=/trunk/kdepim/; revision=299913