This is one we might argue about from a usability perspective.
In the incoming messages dialog, disable buttons that are not
applicable (or do nothing) in the current context. That means:

* Disable accept all when there are no invitations.
* Disable accept and reject when there are no invitations selected.

I think it's best to indicate to the user which actions make
sense.  I'm not sure if I covered all cases where the selection
changes and the buttons might need to be re-configured.

CCMAIL: kde-pim@kde.org

svn path=/trunk/kdepim/; revision=307776