If you copy a todo to the clipboard, don't delete the recurrence (it was removed from the original todo, too)! Also, if you copy anything, don't remove the selection in the todo views, which is inconsistent with the rest of KDE (and just confusing). If the user presses Ctrl+V and then Ctrl+C, okay, he inserts the todo as a subtodo.  But at least the display doesn't change when copying to the clipboard.
Bram, I hope you don't mind about that last change. The first was certainly a bug.

Also add code to delete only the future occurences of a recurring incidence. The code is there, but I still need to find a four-button message box to present the user with the options:
-) Delete only this occurence
-) Delete all occurences
-) Delete only future occurences
-) Cancel

CCMAIL: Bram Schoenmakers <bramschoenmakers@kde.nl>

svn path=/trunk/kdepim/; revision=322148