Move all methods that return some kind of textual representation of incidences to a helper class IncidenceFormatter, which has several static methods for incidence tool tips, (pure text) mail bodys for invitations, html representations of invitations (for kmail's bodypart formatter), and for showing a rich-text version in the event viewer.
I also completely rewrote the formatICal method, which is used by kmail's bodypart formatter, since the old version was parsing the invitation as an iCalendar file (which completely disregards all group scheduling fields like the method), and then needed to parse the mail a second time to find the method. This can be easier done by the parseScheduleMessage. Also, there were so many places, which explicitely checked incidence->type()=="Event". I changed all these to visitors, so there are now about six different visitors used by the IncidenceFormatter. The way to get the URL for the links is implemented via a helper class. See the bodypartformatter, or the formatICal DCOP method of korganizer (which I don't want to remove, because korganizer was released with this DCOP method, and applications removing dcop functions can be really a pain as we experienced with kpilot. So I won't do it myself now with korganizer...)

If anything (in particular kmail's body part formatter for invitations) breaks by this commit, please drop me an email (or directly confront me at aKademy).

CCMAIL: kde-pim@kde.org

svn path=/trunk/kdepim/; revision=338983