Forward port of:
CVS commit by tilladam:

Make it impossible to specify a weekly recurrence without any day set.
While the ical standard allows this, strictly speaking, it leads to an
endless loop when setting todos completed which have such a recurrence,
since the assumption that there will be a day is pervasive.

svn path=/trunk/kdepim/; revision=408997