Finally finishe the GUI for the advanced reminder management. This code has partially been in SVN for quite a while, now it's finished. Also, libkcal (and korgac) already had all that functionality, there was just no proper GUI to access it: 1) Multiple reminders per event 2) Reminders that are before/after the start/end of the event 3) Repeating Reminders
This also fixes several usability bugs in the old reminder widgets (e.g. it wasn't properly visible if a sound was set or not, which sound was set, etc. See the usability reports...) And it finishes our second-most wanted issue.

GUI:
Now, before someone starts screaming that this just makes everything more complicated: The ability to add one simple reminder using just the offset is still there in the editor dlg, there's just an additional "Advanced..." button next to it that gives access to these advanced reminder settings.

FEATURE:16493
FEATURE:82105

svn path=/branches/KDE/3.5/kdepim/; revision=443667