A big commit that contains many fixes to the KOrganizer printing subsystem as well as two new features: incidence printing and  yearly printing.
I'm sure there are several bugs we can close with this commit
but I don't have time to hunt them down at this time... the strings
commit deadline is less than 5hrs away!

FEATURE:
GUI:

This commit merges all the following from the kdepim-3.5.5+ branch:

-->SVN commit 609992 and 609993 by kainhofe:

1) Change some calls from x,y,w,h arguments to one const QRect& argument
2) slightly larger border for all the boxes
3) move the week numbers in the month view one pixel to the left (was too close to the box)

4) Add TODO comments about several things that need to be fixed
5) Fix the month view to not print out hundreds of warning of a failed visitor (which isn't true anyway)

-->SVN commit 610230 by kainhofe:

A lot of printing changed. Basically, a complete cleanup of the print system:
1) Instead of x,y,w,y everything works on QRect now
2) the printer object is generated only when needed, not on startup
3) Better borders and spacings around the boxes
4) correctly color the non-working days in the timetable style
5) correctly set and reset the brushes/pens before and after each drawing
6) better positioning of the text on the incidences
7) One centralized method to draw incidence boxes
8) Got rid of the nasty CalPrintHelper class, which was not needed anyway. Instead the methods are now inside CalPrintPluginBase
9) Added a way to get the list of currently selected incidences (Incidence::List CalPrintPluginBase::mSelectedIncidences;) which is accessible inside each plugin's print method.
etc....

-->SVN commit 610690 by winterz:

Framework for new print incidence feature.
Not much here yet.

-->SVN commit 610892 by winterz:

more configurable stuff for the incidence print

-->SVN commit 610894 by winterz:

cleanup some accelerators.

-->SVN commit 610910 by winterz:

removed acceletors using kdesdk/scripts/fixuifiles

-->SVN commit 610915 by winterz:

Incidence printing actually starts to do something useful.
Still a long way to go, but you can now select an event (in monthview, eg),
then click print, then print incidence.. and you get somewhat useful output.

-->SVN commit 611410 by winterz:

Add a print selection to the incidence context menu.
So, now you can RMB menu over an incidence and select Print
and the print dialog comes up.

-->SVN commit 611426 by winterz:

the print selection doesn't belong in the editonly actions.
this way, printing is enabled even for readonly incidences.

-->SVN commit 611503 by winterz:

avoid leakage.

-->SVN commit 611509 by kainhofe:

Started a Year print style to print the whole year on a page. I'll soon need this for the choir planning of next year..

-->SVN commit 611568 by kainhofe:

-) Further implementation of the yearl print style: Now the grid of days
is already printed, non-working days are shaded a bit, and the day
numbers are printed. No events are shown yet.
-) Factor out the calls to draw the subheaders (weekdays in the
weekstyle and month names in the yearstyle)
-) Add a virtual int sortID(); method to the plugin base class. That
value will be used (not implemented yet!) to identify the config pages
in the widget stack and to sort the radio buttons in the style
selection part of the print dialog.

-->SVN commit 611589 by kainhofe:

use the sortID to show the plugins in the given order. Also connect the sortID to the enum of the type (which now holds explicit int values).

-->SVN commit 611803 by kainhofe:

Some more work on the yearly print style: Now the non-recurring single-day events are already printed (TODO: shorten the summaries of long entries, and use a smaller font). Multi-day events and recurring events are yet to be done. The infrastructure for both is already there, but the final touches are missing

-->SVN commit 611906 by kainhofe:

Practically finished the yearly print style. Some boxes / lines are still off by 1 pixel, and the line spacing of the texts is too large. Also, the summaries are printed out in full, but they should be shortened to show also the second and third event on a day. 
Apart from that, everything should work fine.

-->SVN commit 611925 by kainhofe:

Copy'n'Paste error. Write the correct config key to the file so the orienation is really preserved

-->SVN commit 611926 by kainhofe:

Add TODO lines for the Incidence printing style

-->SVN commit 612105 by kainhofe:

Add helper functions to mask the KOrg::CoreHelper object. Now the plugin does no longer need to know that the callback functions are available via the CoreHelper, but instead can directly call the helper fucntions of the CalPrintPluginBase class

-->SVN commit 612106 by kainhofe:

Check for 0-pointer

-->SVN commit 612109 by kainhofe:

If no incidence pointer is given, use the gray brush instead of the transparent default

-->SVN commit 612227 by kainhofe:

Some pixel-adjusting of the yearly print style

-->SVN commit 612495 by kainhofe:

Some cleanup, properly resetting brushes, etc.

-->SVN commit 612526 by kainhofe:

Don't use QPainter::save/restore when a simple setPen(oldpen) also suffices; Some more pixel-adjusting; Use lighter gray in the printouts

-->SVN commit 612818 by winterz:

make header title string consistent with the week prints.

-->SVN commit 613251 by winterz:

Each plugin now provides a test to decide if it should be enabled in
the print dialog.

-->SVN commit 613719 by kainhofe:

Finally implement the most basic stuff of the incidence printing style. The whole date/time/recurrence/reminder box is missing, attachments are not listed yet, subitems are not printed, but apart from that everything works. Yeah, not a lot of stuff left that's working, but it still looks quite decent already



svn path=/branches/KDE/3.5/kdepim/; revision=614012