Forward port of Several commits, in particular SVN commit 632261 by kainhofe:
Initialize variable to shut up the compiler warning.

Forward port of SVN commit 622165 by kainhofe:

Add the ability to the CalendarResource to set a parent widget for
dialogs like the resource selection dialog. This is used in KOrganizer
to make sure that the resource selection dialog is a child of the
editor dialog (and thus disables the editor dialog).

This also fixes the crash described in bug 123350.

svn path=/trunk/KDE/kdepim/; revision=632265