FEATURE: (contracted in prokde35) Implemented the ability to drag the current meeting time in the freebusy view, in order to visually move the meeting to a spot where the attendees are free. r668059 in enterprise branch.
It's now possible to resize the magenta rectangle that represents the event in the freebusy view,
to change its start time, end time and duration (not only move it around, which preserved duration).
r668997 in enterprise branch.

NOTE: It might interfer with the task-linking feature in the canvas (added for kplato 1.x),
but kplato is now switching to kdgantt2 anyway.

svn path=/trunk/KDE/kdepim/; revision=669024