Icon naming spec compliance (code changes - KDE/): edit-add -> list-add
You have no idea how hard it was to reduce the number of
"add" icons from two to one. I nearly died in the process :-S
The difficult part, of course, was to find the answer
to the everlasting question "What does each icon mean?"

And these are my findings:

* If you add anything to a list (or something else that is
  some kind of collection of items), you use list-add.
* If you remove something from that list (or collection)
  without destroying the removed object itself (because
  it's only a reference to something outside),
  you use list-remove.
* If you delete something from that list (or collection)
  and the item that you deleted had its home in the list
  and is therefore destroyed on removing,
  then you use edit-delete.

Phew. Who wants to open up a TechBase article on using icons?

svn path=/trunk/KDE/kdepim/; revision=753984