o Use of Qt style sheets to define appearance of month view cells
o Month have now alternating colors (computed automatically from the colors
  defined in the configuration). This makes it easier to see what's going on
  while scrolling.

o The month view is now scrollable. Four buttons are provided to go back and
  forward one week or one month. The mouse wheel works too. Going back and
  forward one month using PgUp and PgDown works too.

o Added an option in the configuration to turn on/off the display of icons
  in the month view (can save some space, if the icon is not so important).

o Some optimizations, mainly related to scrolling the month view.
  - removed the hash mDateToCell. It is about 3 times faster to compute
    the index in the mCells array using the start date and the date we
    want to get cell from. getCell() does this now.

o ToDo's with only a due date (but no due-time) were not displayed
  (because KDateTime is invalid in such a case, altough KDateTime::date()
   returns a valid date). This is fixed now.

o When clicking on the background of a month view cell, the cell was not
  selected. This is fixed now.

CCBUG:81689
CCBUG:150067

FEATURE:
GUI:


svn path=/trunk/KDE/kdepim/; revision=772912