Backport r960737 by smartins from trunk to the 4.2 branch:
Fix splitter sizes when using decoration plugins.

When going to configuration dialog and enabling Date Numbers Plugin, with "show on top", it would be shown at bottom, and only shown
correctly when korganizer was restarted.

This had the side effect that when korganizer was closed, it stored in korganizerrc splitter sizes with the third value belonging to
the incorrectly shown decoration, the next time korganizer was started it used the third value for the agenda.

svn path=/branches/KDE/4.2/kdepim/; revision=960982