A few fixes for resource colouring:
1. Colors on subresources weren't being shown sometimes ( when treeWidget->visualItemRect() returned 0 ).

2. Colors were disabled when agenda view settings only used category colors, now monthview settings are respected too.

--//--

I found that "disable colors" doesn't work properly, when korganizer starts colors are enabled again, this is because (i think):

a) "Disable colors" just sets an invalid QColor, but it's still written has an r, g, b set to config, when korganizer starts
   it will read the rgb values into a valid QColor, items will display it as it's valid now.

b) KOPrefs:resourceColor() will fallback to a default color when the color is invalid anyway.


Is this worth fixing or should we remove the "Disable Color" button? It's just there to disable the little colored square, agenda and 
month views always use coloring regardless of colors being or not disabled in resource view.

svn path=/trunk/KDE/kdepim/; revision=961855