Fix segfaults when history entries contain to-dos with references to deleted children.
The crash happens because a clone()d parent still has relations. If you have Parent1 with Child1, and do a Parent1.clone(), the clone will think
that Child1 is his child, but the child doesn't know the clone, so will never notify it when it gets deleted.

This could be fixed in KCal, i think either the clone has 0 relations or we clone the whole family recursively, currently we have an inbetween approach.



svn path=/trunk/KDE/kdepim/; revision=1129520