Fixes for encoding problems in VCards, notably:
kolab/issue3548: D'n'D a contact with umlauts, doesn't create a wrong "to"-header in a mail.
kolab/issue3548: D'n'D a contact with umlauts to Calendar or Todo part creates attendees with wrong encoding in their name
issue3320: Export a contact with umlauts in his name as VCard3 and import it, breaks the umlaut
issue4204: Copying a contact to different resource breaks the umlauts.
issue4054: Umlauts and sending contacts problem(rt#5956)

MERGE: none

svn path=/branches/kdepim/enterprise/kdepim/; revision=1135169