Make the "open" action do the same as "import".
As talked in the last meeting, support for opening a separate window with an individual
ics file was removed and everything must go through akonadi.

Some paths were still opening the window, which lead to a crash.

Now when you "open", you'll be asked if you want to merge or add as new calendar.