allow configuring first day of week. BUG: 358106 FIXED-IN: 5.8.0
note that we've only supported this in the agenda and month views.
it is currently not available in other views, like timeline or whatsnext.
open new wishes if you need those views supported.

also note that you need to restart Kontact/KOrganizer to see the change;
which is yet-another-bug. (in agendaview, moving the range also works)