korgac: Add an option to not grab keyboard focus when a reminder is displayed
Summary:
While it can make it easier to dismiss (if it grabs focus through the
activateWindow call), it also means if you are typing and a reminder
comes in it will both get those keystrokes and a space or return will
dismiss the reminder before you have a chance to ever see what the
reminder was about, which assumes you are even looking at your screen
to see the reminder box flash up, if you are typing in a paper you
might not even know.

Test Plan:
Setup two notifications in korganizer one minute apart.
I the dock window menu test once with "Reminder grabs focus" enabled and once with it disabled.
Give focus to some other program before the first alarm and second alarm, don't dismiss the reminder box between alarms.
With "Reminder grabs focus" checked, the Reminders window should get focus on the second reminder, (this is also the behavior before this change).
With "Reminder grabs focus" unchecked, the Reminders window should not get focus on the second reminder.

Note I'm writing this up for two reminders to avoid the case of window managers which automatically give focus on every new window that pops up.  That's a policy choice on the user/window manager and if this did something to not get focus in that case, this would be going against that policy.

Reviewers: mlaurent, dvratil

Reviewed By: dvratil

Subscribers: dvratil, kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D16242