Patch from David Fries
Summary

Instead of saving position sometimes (application close, suspend
reminder), save at each point the dialog hides, dismiss reminder,
dismiss all, escape or window close, etc.

In addition to saving the position, save the geometry to get the size
as well, that way if the user wants it bigger (or smaller) the only
have to do it once.

The logic was show, move, which would run in that order, the window
shows, then it moves to the correct location which looks bad, restore
the position & size, then show, so it appears in the final location.

BUG: 347486
Test Plan

A quicker way to test this is with a previous patch "Show Reminders" dock menu item option, as opposed to setting a bunch of reminders and waiting for them to go off.

before patch is applied/after patch is applied behaviors

    move the Reminder window to a corner of the screen, show the Reminder window [ before/after it will appear at the same location]
    move the Reminder window to a different corner, press escape, show the Reminder window [ before the escape position didn't save and it appears in the previous location, after it appears in the escape position]
    make the Reminders window large, show Reminders window [ before/after it is the large size]
    exit korgac, run korgac, show Reminders window [ before it is back to the previous normal size, after it is the larger size]

Reviewers: mlaurent, dvratil

Reviewed By: dvratil

Subscribers: dvratil, kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D16245