korgac: Call setOnAllDesktops from showEvent
Earlier, possibly Qt4 versions of korgac had the behavior of once the window manager flagged the window as sticky to show on all desktops, the setting would stay for the life of korgac, but with Qt5 that is cleared on every hide/show.

Test Plan:
Verify the Reminder window is flagged to show on all desktops automatically.

Differential Revision: https://phabricator.kde.org/D16247