Show all available holiday regions
Summary:
The previous approach was forcing unique labels, which fails in many cases
where we have multiple holiday files for a country. In most cases their
disambiguation is only in the description, not the (country) name.

This would be nicer if name() would actually contain the full region name,
not just the country, but lacking that I didn't see an alternative to
squeezing the description into the combo box as well.

As a result I can now actually pick from about a dozen holiday regions for
Germany for example, rather than getting a random one of those.

Reviewers: dvratil

Reviewed By: dvratil

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D29416