korgac: on first run, show alarms from the last 10 days.
Summary:
This was already the intent, but it was broken, nothing showed up.
"from" was an invalid date, and it was passed as remindAtDate,
so the removed if() was always false.

I don't see what sense it makes to pass the "previous check date" as
remindAtDate anyway, so now I'm passing "now" as date. But that seems
redundant with item->mTrigger....

Anyhow for me this fixes the main bug: reminders show up for a fresh
user (with existing calendars).

Test Plan:
killall korgac; remove the CalendarsLastChecked entry from ~/.config/korgacrc
and restart korgac, the last 10 days of reminders should show up.

Reviewers: winterz

Reviewed By: winterz

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D29360