Show occurrence details in Event Viewer
If the user selects an occurrence of a recurring item in the Agenda or
Month views and the Item Viewer is visible in the sidebar, it shows
the start date (for events) or start and due dates (for to-dos) of that
occurrence.  The Event Viewer created from the right-click menu does
something less useful:  it displays the start date of the _first_
occurrence and the due date of the _first incomplete undissociated_
occurrence, no matter which occurrence is clicked.

This patch passes the occurrence's date down to the Event Viewer to
specify which occurrence to display.