Offer the usual choices when deleting recurring to-dos
When the user deletes or drags a recurring event, KOganizer offers to
act on the selected recurrence only, on it and future recurrences, or on
all recurrences.  KOrganizer offers the same choices when the user drags
a recurring to-do, but not when deleting a recurring to-do.

This patch uses the same flow when deleting events and to-dos.  Note that
the choice between deleting child to-dos and making them independent is
only useful when deleting all recurrences;  this patch displays a second
dialog instead of adding another choice to the main dialog.