Remove support for email alarms
That is, sending an email with receivers and content controlled by the iCal
alarm. This is almost as dangerous as procedure alarms in the context of
shared calendars, as it allows an external party to send emails from your
system.

Email alarms are now treated just like regular alarms.

KAlarm still implements this functionality in combination with safe local
calendars, in case someone really can't live without it.