Today button moves Month view to today's month
If the date navigator has selected a week or a work week, the `Today`
button switches the Month view to the month containing the first day of
the selected period containing "today".  On April 2 2022 it showed
March, because the week started on March 28.

This patch uses the "preferred month" infrastructure.  On March 31 it
would show March, and on April 1 it would show April.

BUG: 333066