Toggle completion of occurrences of recurring todos
Currently the 'Toggle To-do Completed' action of the month and agenda
views, applied to a recurring todo, marks the next two incomplete
occurrences as complete, regardless of which occurrence was selected.

This patch calculates the selected occurrence's recurrence ID, and
toggles that occurrence.  If it was incomplete, it and all previous
occurrences are complete;  if it was complete, it and all following
occurrences are incomplete.