Adapt code to GenerateProperties=true in CalendarSettings
This change the return type the <Type>Item from
KConfigSkeletonGenericItem<Type> to KConfigCompilerSignallingItem.

Which are similar but not quite the same (the
KConfigCompilerSignallingItem is a wrapper around the GenericItem)
so add an overload for that in addWidBool.

See https://invent.kde.org/pim/akonadi-calendar/-/merge_requests/60


(cherry picked from commit 04ff7956076ccafaf453173e934698e845403ff5)