Reuse CollectionCalendar instances between views
The goal is to preserve resources and also avoid single Incidence
being member of multiple KCalendarCore::Calendars, which doesn't
work very well.

Right now all incidences are "owned" by the ETMCalendar at the
beginning but are eventualy reparented to the CollectionCalendars
created on demand for views.

Eventually, if we remove ETMCalendar, this problem will go away.