Fix incidence view not updating when incidence is re-selected
When user clicks on an incidence in the event view, the incidence
gets displayed in the incidence view in the left pane. If the user
then clicks on some empty space, the incidence view is cleared, but
CalendarView::processIncidenceSelection() wouldn't clear
mSelectedIncidence. If the user then clicks on the same incidence as
before, the incidence view would not get updated, since the initial
check that selectedItem != mSelectedIncidence would be false, so the
function returned early without updating the incidence view.