Fix crash when forwarding incidence via context menu
The MailClient inside ITIPHandler may work asynchronously, causing the
ITIPHandler inside the KOEVentPopupMenu to go out-of-scope and be
destroyed before the async operation in the MailClient finishes. Since
MailClient relies on a factory object owned by the ITIPHandler, if the
ITIPHandler gets destroyed before the MailClient finishes, it will
delete the factory object causing use-after-free when the async
operation in MailClient is finally done.

BUG: 474144
FIXED-IN: 23.08.3