Use correct timezone when creating recurrence exception
When an incidence occurs on a different day in its original timezone
than in the user's local timezone, deleting a recurrence gives us
a date in the user's local timezone. However, in order for the
recurrence exception to work, the exception date must fall on a date
in the incidence's original timezone.

This change makes sure that the recurrence exception is correctly
converted from local time to the incidence's original timezone before
storing it.

This commit only covers deleting a recurrence ocurrence from KOrganizer
event view. Additional similar patch for the Incidence Editor is needed
to show/create exceptions correctly there.

BUG: 481305
FIXED-IN: 24.02