Remove the 'Custom Pages' settings from KOrganizer
While the KCM somewhat works, it's not implemented in the
incidence editor, so whatever the users manages to set up
in the KCM, it will never work. As per the discussion on
the kde-pim mailing list, let's remove the feature as it
has been broken for years and no-one complained, so it also
likely wasn't used.