Remove options from KOPrefs that already exist in CalendarSupport::KCalPrefs
Some of the holiday-related options were duplicated in both KOPrefs (KOrganizer
config) and KCalPrefs (CalendarSupport config used by other components like
EventViews). While both objects operate on top of the same config file, if the
holidays option got changed through one of the objects, the other one was unaware
of the change.

This caused a weird behavior where re-configuring holidays in KOrganizer showed
up only where KOPrefs object was used to read the configuration, but wouldn't
show up in event views, which use KCalPrefs until the application was
restarted.

BUG: 483504
FIXED-IN: 24.05