akonadicollectionview.cpp - add more info to the delete folder dialog
In the delete calendar dialog, mention that no data
or calendar is deleted at all. and that the calendar can
be added back at a later date.

In the delete folder dialog, give a big warning that
data will be removed for good.


(cherry picked from commit 843283247744728abcf4bff599229b323d9d2367)

Co-authored-by: Allen Winter <winter@kde.org>