Clarify "Working Schedule" vs. "Weekly Schedule"
Some of us don't work.  But we all have what we might consider
to be a "daily schedule", which would exclude weekends.

Try to reduce the emphasis on the concept of a "work week" and
transition to the concept of a "schedule".

This might help relive some bugs where people are disabling
the Working schedule since they don't work (like me).