akonadicollectionview.cpp - quiet the "no items found" popups on delete
When deleting an agent tell the incidencechanger to be quiet
and not display a popup dialog for each and every incidence.

Deleting an agent always succeeds for valid collections and
there is no such thing as deleting in this situation -- all
we're doing is detaching the agent.

BUG: 508530


(cherry picked from commit ea1fa686fa6c360233b9a243d283a11ac57d8d14)

Co-authored-by: Allen Winter <winter@kde.org>