Fix KMime API use
Headers::from7BitString() expects the header content only (not including
the header type and the separating colon), Headers::as7BitString(true)
(true being the default argument) however includes the leading header
type.

That's an API trap, which is why the bool argument for as7BitString()
is going away eventually, but for that we first need to normalize all
uses to not rely on having the header type included anymore.