monthview - no longer show the month/year header
The Monthview month/year header is redundant and takes vertical space.
The month and year is shown in the navigator; if the Monthview is in
fullwindow mode, the month/year header is shown by the horizontal
monthview controls.

A month/year header will not be shown if the user diables the
navigator and monthview is in fullwindow mode,